Update FinalProject.R
This adds all of the raw data into one data frame called "data". There is over 1.2Million rows in this data table, but my code normalized all of it, so nothing should be incorrect, and the only NA's are from the actual data being missing from the LA metro website.