Update FinalProject.R
I added the code that makes the list of rentals turn into hourly demand. We can start forecasting now once we have a training and testing set. We WILL run into a problem though. You cannot compute MAPE with zero's in your data, so this is what professor was talking about in his email.